requirements.txt
# MLOps & Tracking
mlflow==3.9.0
dagshub==0.6.5
wandb==0.24.2
zenml==0.93.2
fastapi==0.128.7
uvicorn==0.40.0
python-dotenv==1.2.1
gradio==6.5.1

# Core Dependencies
numpy==1.25.2
pandas==1.5.3
matplotlib==3.6.0
seaborn==0.13.2
tf_keras==2.20.1

# Data Science & AutoML
scikit-learn==1.2.2
xgboost==3.1.3
TPOT==1.1.0
autogluon==1.5.0
FLAML==2.5.0
h2o==3.46.0.9
pycaret==3.2.0
lightautoml==0.2.7
# auto-sklearn  # Note: Usually requires Linux/WSL
optuna==4.7.0
shap==0.49.1
lime==0.2.0.1
evidently==0.7.20
scipy==1.10.1

# Export & Serving
onnx==1.20.1
ml_dtypes==0.5.4
onnxruntime==1.24.1
skl2onnx==1.20.0
Flask==3.1.2

# Computer Vision & NLP
ultralytics==8.4.14
opencv-python==4.11.0.86
torchvision==0.24.1
transformers==4.57.6
datasets==2.2.1
scipy==1.10.1
pillow==11.3.0
prophet==1.3.0
accelerate==1.12.0
evaluate==0.4.6

# Monitoring & Testing
pytest==9.0.2
httpx==0.28.1

# Time Series
prophet==1.3.0
